Allow license URLs to be customised by calling code.
Currently the license acquisition URLs for protected content (e.g. PlayReady) are baked into the media files, and there isn't a way to change this at run-time. Have added a way for the client code to set an (optional) resolver for these URLs based on the keySystem identifier (e.g. com.microsoft.playready)